I can answer first half of the problem now:
In the computation of the tfidf data, distance metric 'madlib.dist_angle' is used,
which is defined as the angle between two vectors. And it is calculated as Arccos
of the inner product of the two vectors divided by the product of the norms of them.
However, when either one of the two vectors has a norm of 0, we will have Arccos(Inf),
which is carried to be the weight and put into the Boost function of Bernoulli. And
at this point, it fails the assertion of p >= 0.

How the system deals with Inf depends on g++ and the system itself. I guess this is
why the compiled version of MADlib behaves differently on DCA.

Actually, when the norm is zero, the angle is not defined mathematically. Any value
is valid.

Simple fix: in the file of src/modules/linalg/metric.cpp, let the function of
distAngle return a finite value between 0 and pi when either one of the two
norms is zero.