Improvement on the backtracking stepsize while editing the design doc
The original backtracking starts in each iteration from the step size of
the previous iteation.

My previous implementation starts in each iteration from the max step
size.

This commit starts the search for step size from the previous step size
times eta. This is an adaptive way of changing the step size. Some tests
show this is the fastest.